refactor: useForm 훅 추출 및 유틸 함수 추가
