+++ src/types/utilType.ts
@@ -1,0 +1,5 @@
+export type ElementsWith<T> = HTMLFormControlsCollection & {
+  [K in keyof T]: {
+    value: T[K];
+  };
+};
